+++ pandera/core/__init__.py
@@ -11,7 +11,7 @@ from pandera.backends.pandas.hypotheses import PandasHypothesisBackend
 
 from pandera.core.checks import Check
 from pandera.core.hypotheses import Hypothesis
-from pandera.core.pandas import checks, hypotheses
+from pandera.core.pandas import builtin_checks, hypotheses
 from pandera.core.pandas.array import SeriesSchema
 from pandera.core.pandas.components import Column, Index, MultiIndex
 from pandera.core.pandas.container import DataFrameSchema
