+++ pandera/core/pandas/__init__.py
@@ -1,6 +1,6 @@
 """Pandas core."""
 
-from pandera.core.pandas import checks
+from pandera.core.pandas import builtin_checks
 from pandera.core.pandas.array import SeriesSchema
 from pandera.core.pandas.components import Column, Index, MultiIndex
 from pandera.core.pandas.container import DataFrameSchema
