+++ pandera/core/pandas/hypotheses.py
@@ -4,7 +4,7 @@ from typing import Tuple
 
 from pandera.backends.pandas.hypotheses import HAS_SCIPY
 from pandera.core.extensions import register_hypothesis
-from pandera.core.pandas.checks import PandasData
+from pandera.core.pandas.builtin_checks import PandasData
 
 
 if HAS_SCIPY:
